% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swing_points.R
\name{add_swing_high_low}
\alias{add_swing_high_low}
\title{Swing high swing low}
\usage{
add_swing_high_low(x, n = 7)
}
\arguments{
\item{x}{an xts object that contains OHLC data}

\item{n}{Number of periods. Default is 7.}
}
\value{
Adds swing points to a chartSeries
}
\description{
This indicator adds the swing high and swing low points on a chartSeries.
 
The indicator checks historical candles to detect whether there was a 
swing point. This will work by inputting a number to select the range of historical
candles to check to see if there is a swing point in the selected trading period. 
If a candle has the lowest low or highest high in the selected range, then we 
know it was a swing point.
}
\examples{
\dontrun{
library(quantmod)
ADM <- getSymbols("ADM", 
                  from = "2018-01-01",
                  to = "2018-07-01",
                  auto.assign = FALSE)
chartSeries(ADM)
add_swing_high_low(ADM, n = 1)
}
}
\references{
swing high: \url{https://www.investopedia.com/terms/s/swinghigh.asp}
swing low: \url{https://www.investopedia.com/terms/s/swinglow.asp}
}
