% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_marker.R
\name{alert_marker}
\alias{alert_marker}
\title{Alert Marker}
\usage{
alert_marker(x, n = 13)
}
\arguments{
\item{x}{an xts object that contains OHLC data.}

\item{n}{Number of periods for the ema period. Default is 13.}
}
\value{
returns an xts object with marker values.
}
\description{
The alert marker was developed by Alexander Elder. The alert marker is part 
of the impulse system described in his book "Come Into My Trading Room" on 
pages 158-162.
}
\examples{
\dontrun{
library(quantmod)
ADM <- getSymbols("ADM", 
                  from = "2018-01-01", 
                  to = "2018-07-01", 
                  auto.assign = FALSE)
                  
chartSeries(ADM) 
markers <- alert_marker(ADM)

# define colours
cols <- ifelse(markers == quantmod::Hi(ADM), "green", "red")

addPoints(1:nrow(markers), markers, col= cols, pch=19, cex=0.5, on = 1)
}
}
\seealso{
Other Alexander Elder functions: 
\code{\link{safezone}()},
\code{\link{thermometer}()}
}
\concept{Alexander Elder functions}
