% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protective_stops.R
\name{ATR_stop}
\alias{ATR_stop}
\title{Average True Range Trailing Stop}
\usage{
ATR_stop(x, n = 5, coef = 3.5)
}
\arguments{
\item{x}{an xts object that contains OHLC data}

\item{n}{Number of lookback periods for the ATR Stop. Default is 5}

\item{coef}{ATR coefficient. Default is 3.5}
}
\value{
Returns an average true range trailing stop which can be used as a stop loss in a trend following system.
}
\description{
This average true range trailing stop was developed by Sylvain Vervoort.
}
\details{
Based on \href{https://stackoverflow.com/questions/5554220/r-statistics-average-true-range-trailing-stop-indicator}{this question} 
on stackoverflow and \href{http://marintrading.com/106VERV.PDF}{this article} by Sylvain Vervoort.
}
\examples{
\dontrun{
# show ATR stop on chart
getSymbols("ADM", from = "2018-01-01", to = "2018-07-01")
chartSeries(ADM)
addTA(ATR_stop(ADM), on = 1)
}
}
