% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stops.R
\name{chandelier}
\alias{chandelier}
\title{Chandelier Exit}
\usage{
chandelier(HLC, n = 22, coef = 3, trend = "up")
}
\arguments{
\item{HLC}{Object that is coercible to xts or matrix and contains High-Low-Close prices.}

\item{n}{Number of periods for chandelier period. Default is 22.}

\item{coef}{ATR coefficient. Default is 3}

\item{trend}{Indicates if chandelier should be calculated for the up-trend or down-trend. Default is up.}
}
\value{
Returns the chandelier exit points which can be used as stop loss in a trend following strategy.
}
\description{
Calculates the chandelier exits in an up or down trend. The calculated exit 
points are used as a stop loss in a trend following trading strategy.
}
\details{
In an up-trend the default formula is:
Highest High in last 22 days - 3 * ATR for 22 days

In a down-trend the formula is reversed:
Lowest Low in last 22 days + 3 * ATR for 22 days
}
\examples{
\dontrun{
getSymbols("ADM", from = "2018-01-01", to = "2018-07-01")
chartSeries(ADM)
addTA(chandelier(ADM), on = 1)
}


}
