% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Technical Indicators.R
\name{envelope}
\alias{envelope}
\title{Moving Average Envelope}
\usage{
envelope(data, maType = "EMA", n = 22, p = 2.5, ...)
}
\arguments{
\item{data}{an xts object that contains OHLC data}

\item{maType}{moving average type. Default is "EMA". But any}

\item{n}{Number of periods. Default is 22.}

\item{p}{percentage of moving the upper and lower bounderies away from the 
midpoint of the moving average. Default is 2.5. Needs to be between 0 and 100.}

\item{...}{any other passthrough parameters}
}
\value{
Returns a timeseries with the lower and upper bound and the midpoint of the chosen moving average.
}
\description{
This function creates a moving average envelope.
}
\examples{
\dontrun{
getSymbols("ADM", from = "2018-01-01", to = "2018-07-01")
chartSeries(ADM)
addTA(envelope(ADM), on = 1)
}
}
\references{
\url{https://www.investopedia.com/terms/e/envelope.asp}
}
