% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_key_statistics.R
\name{get_statistics}
\alias{get_statistics}
\title{Get Key Statistics}
\usage{
get_statistics(symbol)
}
\arguments{
\item{symbol}{a stock symbol}
}
\value{
Returns a data.frame containing the key statistics from the statistics page
from Yahoo Finance. Available fields are symbol, statistic_group, statistic and value.
All values are character values are only as good as the data on Yahoo.
}
\description{
This function get the information from the key statistics page from yahoo finance. 
For individual statistics or statistics to calculate with, it is better to use 
`quantmod::getQuote` together with `quantmod::yahooQF`.
}
\examples{
\dontrun{
library(dplyr)
# returns the statistics for AAPL
aapl_stats <- get_statistics("AAPL")

# get all the profitability stats
aapl_stat \%>\% 
  filter(statistic_group == "Profitability")
 
# returns the statistics for AAPL and FB
symbols <- c("AAPL", "FB")
symbol_stats <- get_statistics(symbols)

symbol_stats \%>\% 
  filter(statistic == "Forward Annual Dividend Rate")
}
}
