% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heikin_ashi.R
\name{heikin_ashi}
\alias{heikin_ashi}
\title{Heikin-Ashi}
\usage{
heikin_ashi(data)
}
\arguments{
\item{data}{an xts object that contains OHLC data}
}
\value{
The heikin_ashi function will return a OHLC dataset with the calculated 
heikin ashi values needed for plotting in a candlestick chart.
}
\description{
Calculates the Heikin-Ashi values needed for plotting in a candlestick chart.
}
\details{
1. The Heikin-Ashi Close is an average of the open, high, low and 
 close for the current period. 
 
 2. The Heikin-Ashi Open is the average of the prior Heikin-Ashi open and the 
 prior Heikin-Ashi close
 
 3. The Heikin-Ashi High is the maximum of: 
 the current period's high, the current Heikin-Ashi open or the current 
 Heikin-Ashi close.
 
 4. The Heikin-Ashi low is the minimum : 
 the current period's low, the current Heikin-Ashi open or the current 
 Heikin-Ashi close.
 
 More information can be found at \href{https://www.investopedia.com/trading/heikin-ashi-better-candlestick/}{Investopedia}
}
\examples{
\dontrun{
ADM <- getSymbols("ADM", 
                  from = "2018-01-01", 
                  to = "2018-07-01", 
                  auto.assign = FALSE)
ha_ADM <- heikin_ashi(ADM)
chartSeries(ha_ADM)
}
 
}
