% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protective_stops.R
\name{safezone}
\alias{safezone}
\title{Safezone Stop}
\usage{
safezone(x, n = 10, coef = 2, prevent = 5, trend = "up")
}
\arguments{
\item{x}{An object that is coercible to xts or matrix and contains High-Low prices.}

\item{n}{Number of lookback periods for the safezone period. Default is 10.}

\item{coef}{Coefficient for multiplying the average downside (or upside) penetration. Default is 2.}

\item{prevent}{Prevents the stop from lowering for x days. Default is 5.}

\item{trend}{Indicates if the safezone stop should be calculated for the up-trend or down-trend. 
Default is "up". Possible values are "up" or "down".}
}
\value{
Returns the safezone exit points which can be used as stop loss in a trend following strategy.
}
\description{
The safezone stop was developed by Alexander Elder. The safezone stop is described in his book "Come Into My Trading Room".
}
\details{
Once you have defined a trend, you need to choose the length of the lookback period for measuring noise level. 
It has to be long enough to track recent behavior but short enough to be relevant for current trading.

If the trend is up, mark all downside penetrations during the lookback period, 
add their depths, and divide the sum by the number of penetrations. This gives you
the average downside penetration for the selected lookback period. Multiply 
the average downside penetration by a coefficient, to place stops farther away from the market 
than the average level of noise. The results are subtracted from yesterday's low to create the stop. 
To prevent the stop from lowering too fast, the prevent lookback option will keep the maximum calculated 
stop for the specified lookback. The rules are reversed for down trend calculations.
}
\examples{
\dontrun{
# show safezone stop on chart
getSymbols("ADM", from = "2018-01-01", to = "2018-07-01")
chartSeries(ADM)
addTA(safezone(ADM), on = 1)
}
}
