% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_thermometer.R
\name{thermometer}
\alias{thermometer}
\title{Market Thermometer}
\usage{
thermometer(x, n = 22)
}
\arguments{
\item{x}{an xts object that contains OHLC data.}

\item{n}{Number of periods for the ema period. Default is 22.}
}
\value{
returns an xts object containing the columns:
 \itemize{
   \item temperature contains the calculated temperature values
   \item temp_ema Ema values based on the temperature values.}
}
\description{
The market thermometer was developed by Alexander Elder. The market
thermometer is described in his book "Come Into My Trading Room" on pages 
162-164.
}
\examples{
\dontrun{
library(quantmod)
ADM <- getSymbols("ADM", 
                  from = "2018-01-01", 
                  to = "2018-07-01", 
                  auto.assign = FALSE)
                  
chartSeries(ADM, TA = NULL) # without volume
addTA(thermometer(ADM))
}
}
\seealso{
Other Alexander Elder functions: 
\code{\link{alert_marker}()},
\code{\link{safezone}()}
}
\concept{Alexander Elder functions}
