% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_channel.R
\name{yield_channel}
\alias{yield_channel}
\title{Yield Channel}
\usage{
yield_channel(ticker, n = 12, dividend_channel = TRUE, dividend_bands = TRUE)
}
\arguments{
\item{ticker}{A stock symbol}

\item{n}{number of years to look back. Ideally one or two business cycles. 
Default look back period is 12 years.}

\item{dividend_channel}{Show the dividend channels. Default is TRUE.}

\item{dividend_bands}{Show the dividend bands. Default is TRUE.}
}
\value{
Returns a chart showing the price, the dividend channels and dividend 
bands.
}
\description{
Shows the dividend channels of a dividend stock.
}
\details{
If a dividend stock's price is below the channel it might indicate that it is 
undervalued based on the historical dividend yields. If it is above the 
channel it might indicate that it is overvalued. The dividend bands show bands 
of slightly overvalued to highly overvalued and slightly undervalued and highly
undervalued. 

The calculation for the dividend channels are based on the 5\% and the 95\% value 
of the dividend percentages and smoothed out via a loess regression. The stock 
is undervalued if its current dividend yield is higher than the average dividend 
yield of the past based on the chosen look back period. The shorter the look 
back period, the faster a stock might appear under- or overvalued. 

The dividend bands show ranges between 5\%-25\%, 25\%-50\%, 50\%-75\% and 75\%-95\%
percentage. These bands are not smoothed out and will show the steps in 
dividend increases.

More information can be found in \href{https://seekingalpha.com/article/4393616-revisiting-yield-channel-charts}{this Seeking Alpha} article.
}
\examples{
\dontrun{
# shows the yield channel chart for Merck.
yield_channel("MRK")
}
}
