% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_indicators.R
\name{envelope}
\alias{envelope}
\title{Moving Average Envelope}
\usage{
envelope(x, ma = "EMA", n = 22, p = 2.5, ...)
}
\arguments{
\item{x}{an xts object that contains OHLC data}

\item{ma}{Type of moving average used. Default is "EMA". Options are EMA or SMA.}

\item{n}{Number of periods. Default is 22.}

\item{p}{percentage of moving the upper and lower bounderies away from the 
midpoint of the moving average. Default is 2.5. Needs to be between 0 and 100.}

\item{...}{any other passthrough parameters}
}
\value{
Returns a timeseries with the lower and upper bound and the midpoint of the chosen moving average.
}
\description{
This function creates a moving average envelope.
}
\examples{
\dontrun{
library(quantmod)
ADM <- getSymbols("ADM", 
                  from = "2018-01-01",
                  to = "2018-07-01",
                  auto.assign = FALSE)
chartSeries(ADM)
addTA(envelope(ADM), on = 1)
}
}
\references{
\url{https://www.investopedia.com/terms/e/envelope.asp}
}
